source/dissertation/client/requirements.txt
-r ../shared/requirements.txt

flask-restplus